refactor: include erlc data files in dependant directory
