+++ Source/Commands/Miscellaneous/Log/Deps/CitWarn.ts
@@ -15,9 +15,9 @@ import {
   WeatherConditions,
 } from "@Models/Citation.js";
 
-import ERLCAgeGroups from "@Resources/ERLCAgeGroups.js";
+import ERLCAgeGroups from "@Resources/ERLC-Data/ERLCAgeGroups.js";
 import AnyCitationCallback from "./Funcs/AnyCitationHandler.js";
-import { EyeColors, HairColors } from "@Resources/ERLCPDColors.js";
+import { EyeColors, HairColors } from "@Resources/ERLC-Data/ERLCPDColors.js";
 import { ReporterInfo } from "../Log.js";
 
 const CmdFileLabel = "Commands:Miscellaneous:Log:CitWarn";
