+++ Source/Resources/ERLC-Data/ERLCPDColors.ts
