+++ Source/Typings/Utilities/Database.d.ts
@@ -1,6 +1,6 @@
 import type { IncidentTypes, IncidentStatusesFlattened } from "@Resources/IncidentConstants.ts";
 import type { Types, HydratedDocument, Model } from "mongoose";
-import type { EyeColors, HairColors } from "@Resources/ERLCPDColors.ts";
+import type { EyeColors, HairColors } from "@Resources/ERLC-Data/ERLCPDColors.ts";
 import type { ShiftFlags } from "@Models/Shift.ts";
 import type { Overwrite } from "utility-types";
 import type {
@@ -12,7 +12,7 @@ import type {
   RoadSurfaceConditions,
 } from "@Models/Citation.ts";
 
-import type ERLCAgeGroups from "@Resources/ERLCAgeGroups.ts";
+import type ERLCAgeGroups from "@Resources/ERLC-Data/ERLCAgeGroups.ts";
 import type AppError from "@Utilities/Classes/AppError.ts";
 
 export namespace Guilds {
