+++ Source/Utilities/Strings/Formatters.ts
@@ -6,7 +6,7 @@ import { GuildCitations } from "@Typings/Utilities/Database.js";
 import { userMention } from "discord.js";
 import { TitleCase } from "./Converters.js";
 import { Vehicles } from "@Typings/Resources.js";
-import ERLCAgeGroups from "@Resources/ERLCAgeGroups.js";
+import ERLCAgeGroups from "@Resources/ERLC-Data/ERLCAgeGroups.js";
 import DHumanizer from "humanize-duration";
 import _Dedent from "dedent";
